Add typescript tests
