+++ remix-resolve/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "build": "tsc",
     "lint": "standard",
-    "test": "standard && mocha tests/ -t 300000"
+    "test": "standard && mocha --require ts-node/register tests/*.ts -t 300000"
   },
   "repository": {
     "type": "git",
@@ -35,8 +35,14 @@
     "valid-url": "^1.0.9"
   },
   "devDependencies": {
+    "@types/chai": "^4.1.7",
+    "@types/mocha": "^5.2.5",
+    "@types/node": "^10.12.18",
+    "chai": "^4.2.0",
     "mocha": "^5.1.0",
+    "remix-plugin": "0.0.1-alpha.2",
     "standard": "^12.0.1",
+    "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
     "typescript": "^3.1.6"
   }
